# Gradle 7 upgrade for openhouse oss repo