--- a/apps/spark/build.gradle
+++ b/apps/spark/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'openhouse.hadoop-conventions'
     id 'openhouse.iceberg-conventions'
     id 'openhouse.maven-publish'
-    id 'com.github.johnrengelman.shadow' version '6.0.0'
+    id 'com.github.johnrengelman.shadow' version '7.1.2'
 }
 
 configurations {
@@ -38,7 +38,7 @@ dependencies {
     implementation 'org.apache.iceberg:iceberg-spark-runtime-3.1_2.12:' + icebergVersion
     implementation 'commons-cli:commons-cli:1.5.0'
     implementation 'org.reflections:reflections:0.10.2'
-    implementation 'org.springframework.boot:spring-boot-starter-webflux:2.6.6'
+    implementation 'org.springframework.boot:spring-boot-starter-webflux:2.7.14'
     implementation 'io.netty:netty-resolver-dns-native-macos:4.1.70.Final:osx-x86_64'
     implementation 'org.springframework.retry:spring-retry:1.3.3'
     implementation 'org.apache.logging.log4j:log4j-core:2.18.0'
@@ -67,7 +67,7 @@ dependencies {
     testImplementation 'org.powermock:powermock-module-junit4:2.0.9'
     testImplementation 'org.powermock:powermock-api-mockito2:2.0.9'
     testImplementation(project(':tables-test-fixtures_2.12'))
-    testRuntime("org.eclipse.jetty:jetty-server:11.0.2")
+    testRuntimeOnly("org.eclipse.jetty:jetty-server:11.0.2")
 
 }