--- a/apps/spark/build.gradle
+++ b/apps/spark/build.gradle
@@ -14,6 +14,12 @@ configurations {
     shadow.extendsFrom implementation
 }
 
+configurations.all {
+    resolutionStrategy {
+        force 'org.apache.commons:commons-lang3:3.12.0'
+    }
+}
+
 ext {
     sparkVersion = '3.1.1'
     icebergVersion = '1.2.0'
@@ -27,19 +33,36 @@ dependencies {
     implementation project(':client:secureclient')
     implementation project(':services:common')
     implementation project(':cluster:storage')
-    compileOnly project(':integrations:spark:openhouse-spark-runtime_2.12')
+    compileOnly (project(path: ':integrations:spark:openhouse-spark-runtime_2.12')) {
+        exclude group: 'io.netty'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-common'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-client'
+    }
     implementation ('org.apache.spark:spark-core_2.12:' + sparkVersion) {
         exclude group: 'io.netty'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-common'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-client'
     }
     implementation ('org.apache.spark:spark-sql_2.12:' + sparkVersion) {
         exclude group: 'io.netty'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-common'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-client'
     }
-    implementation 'org.apache.hadoop:hadoop-common:2.10.0'
-    implementation 'org.apache.iceberg:iceberg-spark-runtime-3.1_2.12:' + icebergVersion
+    implementation ('org.apache.hadoop:hadoop-common:2.10.0') {
+        exclude group: 'io.netty'
+        exclude group: 'org.apache.curator', module: 'curator-client'
+        exclude group: 'org.apache.commons', module: 'commons-lang3'
+        
+    }
+    implementation ('org.apache.iceberg:iceberg-spark-runtime-3.1_2.12:' + icebergVersion) {
+        exclude group: 'io.netty'
+    }
     implementation 'commons-cli:commons-cli:1.5.0'
     implementation 'org.reflections:reflections:0.10.2'
-    implementation 'org.springframework.boot:spring-boot-starter-webflux:2.7.14'
-    implementation 'io.netty:netty-resolver-dns-native-macos:4.1.70.Final:osx-x86_64'
+    implementation ('org.springframework.boot:spring-boot-starter-webflux:2.7.8') {
+        //exclude group: 'io.netty'
+    }
+    implementation 'io.netty:netty-resolver-dns-native-macos:4.1.75.Final:osx-x86_64'
     implementation 'org.springframework.retry:spring-retry:1.3.3'
     implementation 'org.apache.logging.log4j:log4j-core:2.18.0'
     implementation 'org.apache.logging.log4j:log4j-slf4j-impl:2.18.0'
@@ -52,6 +75,7 @@ dependencies {
     implementation 'com.fasterxml.woodstox:woodstox-core:6.2.7'
 
     // open telemetry related classed. Latest Okhttp version is 4.10.0, pinning to 4.9.3 to avoid dependency issues
+    implementation(platform("com.squareup.okhttp3:okhttp-bom:4.9.3"))
     implementation 'com.squareup.okhttp3:okhttp:4.9.3'
     implementation 'com.squareup.okio:okio:3.2.0'
     implementation 'io.opentelemetry:opentelemetry-api:1.18.0'
@@ -59,14 +83,24 @@ dependencies {
     implementation 'io.opentelemetry:opentelemetry-sdk:1.18.0'
     implementation 'io.opentelemetry:opentelemetry-sdk-extension-autoconfigure:1.14.0-alpha'
     implementation 'io.opentelemetry:opentelemetry-semconv:1.14.0-alpha'
+    implementation 'org.apache.commons:commons-lang3:3.12.0'
 
-    testImplementation project(':integrations:spark:openhouse-spark-runtime_2.12')
+    testImplementation (project(path: ':integrations:spark:openhouse-spark-runtime_2.12', configuration: 'shadow')) {
+        exclude group: 'io.netty'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-common'
+        exclude group: 'org.apache.hadoop', module: 'hadoop-client'
+    }
     // Otherwise throws the error: Scala module 2.10.0 requires Jackson Databind version >= 2.10.0 and < 2.11.0
     testImplementation 'com.fasterxml.jackson.module:jackson-module-scala_2.12:2.13.1'
     testImplementation 'org.mockito:mockito-inline:4.11.0'
     testImplementation 'org.powermock:powermock-module-junit4:2.0.9'
     testImplementation 'org.powermock:powermock-api-mockito2:2.0.9'
-    testImplementation(project(':tables-test-fixtures_2.12'))
+    testImplementation(project(':tables-test-fixtures_2.12')) {
+        exclude group: "io.netty"
+        //exclude group: 'org.apache.hadoop', module: 'hadoop-common'
+        //exclude group: 'org.apache.hadoop', module: 'hadoop-client'
+    }
+    //testImplementation 'org.apache.hadoop:hadoop-common:3.2.0'
     testRuntimeOnly("org.eclipse.jetty:jetty-server:11.0.2")
 
 }