--- a/apps/spark/build.gradle
+++ b/apps/spark/build.gradle
@@ -60,9 +60,7 @@ dependencies {
     }
     implementation 'commons-cli:commons-cli:1.5.0'
     implementation 'org.reflections:reflections:0.10.2'
-    implementation ('org.springframework.boot:spring-boot-starter-webflux:2.7.8') {
-        //exclude group: 'io.netty'
-    }
+    implementation ('org.springframework.boot:spring-boot-starter-webflux:2.7.8')
     implementation 'io.netty:netty-resolver-dns-native-macos:4.1.75.Final:osx-x86_64'
     implementation 'org.springframework.retry:spring-retry:1.3.3'
     implementation 'org.apache.logging.log4j:log4j-core:2.18.0'
@@ -98,10 +96,7 @@ dependencies {
     testImplementation 'org.powermock:powermock-api-mockito2:2.0.9'
     testImplementation(project(':tables-test-fixtures_2.12')) {
         exclude group: "io.netty"
-        //exclude group: 'org.apache.hadoop', module: 'hadoop-common'
-        //exclude group: 'org.apache.hadoop', module: 'hadoop-client'
     }
-    //testImplementation 'org.apache.hadoop:hadoop-common:3.2.0'
     testRuntimeOnly("org.eclipse.jetty:jetty-server:11.0.2")
 
 }