--- a/apps/spark/build.gradle
+++ b/apps/spark/build.gradle
@@ -74,7 +74,6 @@ dependencies {
     implementation 'com.fasterxml.woodstox:woodstox-core:6.2.7'
 
     // open telemetry related classed. Latest Okhttp version is 4.10.0, pinning to 4.9.3 to avoid dependency issues
-    implementation(platform("com.squareup.okhttp3:okhttp-bom:4.9.3"))
     implementation 'com.squareup.okhttp3:okhttp:4.9.3'
     implementation 'com.squareup.okio:okio:3.2.0'
     implementation 'io.opentelemetry:opentelemetry-api:1.18.0'