--- a/apps/spark/build.gradle
+++ b/apps/spark/build.gradle
@@ -14,12 +14,6 @@ configurations {
     shadow.extendsFrom implementation
 }
 
-configurations.all {
-    resolutionStrategy {
-        force 'org.apache.commons:commons-lang3:3.12.0'
-    }
-}
-
 ext {
     sparkVersion = '3.1.1'
     icebergVersion = '1.2.0'