--- a/build.gradle
+++ b/build.gradle
@@ -26,8 +26,9 @@ buildscript {
 }
 
 ext {
-  spring_web_version = "2.6.6"
+  spring_web_version = "2.7.14"
   spark_version = "3.1.1"
+  springBootVersion = "2.7.14"
 }
 
 group = 'com.linkedin.openhouse'
@@ -126,3 +127,4 @@ tasks.register('CopyGitHooksTask', Copy) {
   from file('scripts/git-hooks')
   into file('.git/hooks/')
 }
+