--- a/build.gradle
+++ b/build.gradle
@@ -26,9 +26,9 @@ buildscript {
 }
 
 ext {
-  spring_web_version = "2.7.14"
+  spring_web_version = "2.7.8"
   spark_version = "3.1.1"
-  springBootVersion = "2.7.14"
+  springBootVersion = "2.7.8"
 }
 
 group = 'com.linkedin.openhouse'