--- a/build.gradle
+++ b/build.gradle
@@ -28,7 +28,6 @@ buildscript {
 ext {
   spring_web_version = "2.7.8"
   spark_version = "3.1.1"
-  springBootVersion = "2.7.8"
 }
 
 group = 'com.linkedin.openhouse'