--- a/buildSrc/src/main/groovy/openhouse.client-codegen-convention.gradle
+++ b/buildSrc/src/main/groovy/openhouse.client-codegen-convention.gradle
@@ -1,6 +1,6 @@
 ext {
   swagger_annotations_version = "1.6.2"
-  spring_web_version = "2.6.6"
+  spring_web_version = "2.7.14"
   jackson_version = "2.11.3"
   jackson_databind_version = "2.11.3"
   jackson_databind_nullable_version = "0.2.1"