--- a/buildSrc/src/main/groovy/openhouse.java-conventions.gradle
+++ b/buildSrc/src/main/groovy/openhouse.java-conventions.gradle
@@ -4,7 +4,7 @@ plugins {
 
 ext {
   gsonVersion = '2.8.9'
-  micrometerVersion = '1.6.1'
+  micrometerVersion = '1.9.13'
 }
 
 dependencies {
@@ -13,5 +13,7 @@ dependencies {
 
   implementation 'io.micrometer:micrometer-core:' + micrometerVersion
   implementation 'io.micrometer:micrometer-registry-prometheus:' + micrometerVersion
+  implementation 'com.squareup.okhttp3:okhttp:4.9.3'
+  implementation 'commons-cli:commons-cli:1.5.0'
 
 }