--- a/buildSrc/src/main/groovy/openhouse.maven-publish.gradle
+++ b/buildSrc/src/main/groovy/openhouse.maven-publish.gradle
@@ -3,17 +3,17 @@ plugins {
 }
 
 task sourcesJar(type: Jar) {
-  classifier = "sources"
+  archiveClassifier = "sources"
   from sourceSets.main.allSource
 }
 
 task javadocJar(type: Jar) {
-  classifier = "javadoc"
+  archiveClassifier = "javadoc"
   from javadoc.destinationDir
 }
 
 task libJar(type: Jar) {
-  classifier = "lib"
+  archiveClassifier = "lib"
   from sourceSets.main.output
 }