--- a/buildSrc/src/main/groovy/openhouse.springboot-ext-conventions.gradle
+++ b/buildSrc/src/main/groovy/openhouse.springboot-ext-conventions.gradle
@@ -1,6 +1,7 @@
 // plugin for extensible springboot services
 plugins {
   id 'openhouse.springboot-conventions'
+  id "org.springframework.boot"
 }
 
 jar {