--- a/buildSrc/src/main/groovy/openhouse.springboot-ext-conventions.gradle
+++ b/buildSrc/src/main/groovy/openhouse.springboot-ext-conventions.gradle
@@ -20,8 +20,8 @@ tasks.named("bootJar") {
 }
 
 configurations {
-  [apiElements, runtimeElements].each {
+  /*[apiElements, runtimeElements].each {
     it.outgoing.artifacts.removeIf { it.buildDependencies.getDependencies(null).contains(jar) }
     it.outgoing.artifact(bootJar)
-  }
+  }*/
 }