--- a/buildSrc/src/main/groovy/openhouse.springboot-ext-conventions.gradle
+++ b/buildSrc/src/main/groovy/openhouse.springboot-ext-conventions.gradle
@@ -18,10 +18,3 @@ tasks.named("bootJar") {
     attributes ('Main-Class': 'org.springframework.boot.loader.PropertiesLauncher')
   }
 }
-
-configurations {
-  /*[apiElements, runtimeElements].each {
-    it.outgoing.artifacts.removeIf { it.buildDependencies.getDependencies(null).contains(jar) }
-    it.outgoing.artifact(bootJar)
-  }*/
-}