--- a/integrations/java/openhouse-java-runtime/build.gradle
+++ b/integrations/java/openhouse-java-runtime/build.gradle
@@ -32,6 +32,7 @@ dependencies {
   compileOnly("org.apache.iceberg:iceberg-spark-runtime-3.1_2.12:" + icebergVersion)
   compileOnly ("org.springframework.boot:spring-boot-starter-webflux:" + springVersion)
 
+  implementation 'org.apache.commons:commons-lang3:3.12.0'
   fatJarPackagedDependencies(project(':client:secureclient'))
   fatJarRuntimeDependencies("org.apache.iceberg:iceberg-core:" + icebergVersion)
   fatJarRuntimeDependencies("org.apache.hadoop:hadoop-client:" + hadoopVersion) {
@@ -39,6 +40,7 @@ dependencies {
     exclude group: 'javax', module: 'servlet-api'
     exclude group: "io.netty"
     exclude group: 'com.zaxxer', module: 'HikariCP-java7'
+    exclude group: 'org.apache.commons', module: 'commons-lang3'
   }
 }