--- a/integrations/spark/openhouse-spark-runtime/build.gradle
+++ b/integrations/spark/openhouse-spark-runtime/build.gradle
@@ -1,7 +1,7 @@
 plugins {
   id 'openhouse.java-minimal-conventions'
   id 'openhouse.maven-publish'
-  id 'com.github.johnrengelman.shadow' version '6.0.0'
+  id 'com.github.johnrengelman.shadow' version '7.1.2'
   id 'scala'
   id 'antlr'
 }