--- a/scripts/java/tools/dummytokens/build.gradle
+++ b/scripts/java/tools/dummytokens/build.gradle
@@ -13,6 +13,7 @@ dependencies {
 }
 
 jar {
+    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
     manifest {
         attributes(
             'Main-Class': 'com.linkedin.openhouse.tools.dummytokens.DummyTokenGenerator'