--- a/services/common/build.gradle
+++ b/services/common/build.gradle
@@ -27,5 +27,5 @@ dependencies {
   testImplementation 'commons-io:commons-io:2.4'
   testImplementation 'org.hamcrest:hamcrest-all:1.3'
   testFixturesImplementation 'commons-io:commons-io:2.4'
-  testFixturesImplementation 'org.springframework.boot:spring-boot-starter-test:2.6.6'
+  testFixturesImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.14'
 }