--- a/services/common/build.gradle
+++ b/services/common/build.gradle
@@ -17,7 +17,7 @@ dependencies {
   implementation 'jakarta.validation:jakarta.validation-api:2.0.2'
   implementation 'org.springframework:spring-web:' + '5.3.18'
   implementation 'org.springframework:spring-webmvc:' + '5.3.18'
-  implementation 'org.springframework.boot:spring-boot-starter-webflux:2.6.6'
+  implementation 'org.springframework.boot:spring-boot-starter-webflux:2.7.8'
   implementation 'org.springframework.security:spring-security-core:' + '5.7.3'
   implementation 'io.jsonwebtoken:jjwt:' + '0.9.1'
   implementation 'org.aspectj:aspectjrt:' + '1.9.19'
@@ -27,5 +27,5 @@ dependencies {
   testImplementation 'commons-io:commons-io:2.4'
   testImplementation 'org.hamcrest:hamcrest-all:1.3'
   testFixturesImplementation 'commons-io:commons-io:2.4'
-  testFixturesImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.14'
+  testFixturesImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.8'
 }