--- a/services/housetables/build.gradle
+++ b/services/housetables/build.gradle
@@ -20,9 +20,6 @@ plugins {
   id 'openhouse.service-specgen-convention'
 }
 
-bootJar {
-  enabled = true
-}
 
 dependencies {
   implementation project(':services:common')