… gradlew
… libs/datalayout/build.gradle
… services/jobs/build.gradle
… services/tables/build.gradle
… tables-test-fixtures/build.gradle