## Summary

<!--- HINT: Replace #nnn with corresponding Issue number, if you are fixing an existing issue -->
This PR performs gradle7 upgrade for the repo. The upgrade is done from gradle 6.9 to 7.6.

Here are the main changes in this PR.
- Gradle 7 specific configuration and deprecation update as per the [gradle 7 upgrade wiki](https://docs.gradle.org/current/userguide/upgrading_version_6.html).
- Although existing spring boot version `2.6.6` is compatible with gradle `7.x`, faced some issues during the upgrade. So upgraded spring boot version to `2.7.8` and this version seems to be working well with gradle `7.x`. Also it would be easier to upgrade gradle `8.x` as the supported spring boot version is `2.7.9` or `2.7.10`. Upgrading directly to `2.7.10` or higher is failing due to old Netty version dependency which are pulled in from hadoop `2.10.0`. Netty version `4.1.86.Final` which is pulled in by spring boot version `2.7.8` is working well with this upgrade whereas higher version of Netty introduces additional issues. Refer to the [spring boot issue](https://github.com/spring-projects/spring-boot/issues/12097) and [Spring 2.7.8 system requirements](https://docs.spring.io/spring-boot/docs/2.7.8/reference/htmlsingle/#getting-started.system-requirements). Also refer to [spring boot 2.7.x dependency versions](https://docs.spring.io/spring-boot/docs/2.7.x/reference/html/dependency-versions.html#appendix.dependency-versions).
- Upgraded spring boot gradle plugin to `2.7.8`
- Upgraded johnrengelman shadow plugin version to `7.1.2`.
- Classifier change for shadow jars. Added `uber` classifier for all the shadow jars.
- Boot jar remains unchanged and no classifier added. Regular jar continues to have `lib` classifier.
- Gradle publication changes to ensure maven publish works.

## Changes

- [ ] Client-facing API Changes
- [ ] Internal API Changes
- [ ] Bug Fixes
- [ ] New Features
- [ ] Performance Improvements
- [ ] Code Style
- [ ] Refactoring
- [ ] Documentation
- [ ] Tests
- [x] Upgrade

For all the boxes checked, please include additional details of the changes made in this pull request.  

## Testing Done
<!--- Check any relevant boxes with "x" -->

- [ ] Manually Tested on local docker setup. Please include commands ran, and their output.
- [ ] Added new tests for the changes made.
- [ ] Updated existing tests to reflect the changes made.
- [ ] No tests added or updated. Please explain why. If unsure, please feel free to ask for help.
- [x] Some other form of testing like staging or soak time in production. Please explain.

- `./gradlew clean build` working and tests are passing. 
- Tested using local docker setup and able to perform table operation after gradle 7 upgrade.
- Ensured that gradle maven publish works by publishing to the local repository.
  ./gradlew publishToMavenLocal -Pversion=1.0.0-snapshot

Ensured that the artifacts produced after the upgrade remains the same and there is no significant increase in size.

**Before upgrade:**
```
anath1@anath1-mn1 openhouse % ls -lrt build/openhouse-java-runtime/libs 
total 81168
-rw-r--r--  1 anath1  101  20766490 Jul 26 12:06 openhouse-java-runtime.jar

anath1@anath1-mn1 openhouse % ls -lrt build/openhouse-spark-runtime_2.12/libs                       
total 81872
-rw-r--r--  1 anath1  101  20931634 Jul 26 12:06 openhouse-spark-runtime_2.12.jar

anath1@anath1-mn1 openhouse % ls -lrt build/tables/libs 
total 561632
-rw-r--r--  1 anath1  101  143504765 Jul 26 12:03 tables.jar
-rw-r--r--  1 anath1  101     214866 Jul 26 12:03 tables-lib.jar

anath1@anath1-mn1 openhouse % ls -lrt build/openhouse-spark-apps_2.12/libs 
total 776136
-rw-r--r--  1 anath1  101     106777 Jul 26 12:06 openhouse-spark-apps_2.12.jar
-rw-r--r--  1 anath1  101  198548473 Jul 26 12:08 openhouse-spark-apps_2.12-uber.jar
```

**After upgrade:**
```
anath1@anath1-mn1 openhouse % ls -lrt build/openhouse-java-runtime/libs                                              
total 41512
-rw-r--r--  1 anath1  101  21093437 Aug  2 10:19 openhouse-java-runtime-uber.jar
-rw-r--r--  1 anath1  101     39568 Aug  2 10:29 openhouse-java-runtime.jar
anath1@anath1-mn1 openhouse % ls -lrt build/openhouse-spark-runtime_2.12/libs                                        
total 43872
-rw-r--r--  1 anath1  101  21258607 Aug  2 10:20 openhouse-spark-runtime_2.12-uber.jar
-rw-r--r--  1 anath1  101    166197 Aug  2 10:21 openhouse-spark-runtime_2.12.jar
anath1@anath1-mn1 openhouse % ls -lrt build/tables/libs 
total 281080
-rw-r--r--  1 anath1  101     214411 Aug  2 10:17 tables-lib.jar
-rw-r--r--  1 anath1  101  143694207 Aug  2 10:26 tables.jar
anath1@anath1-mn1 openhouse % ls -lrt build/openhouse-spark-apps_2.12/libs    
total 426216
-rw-r--r--  1 anath1  101     105922 Aug  2 10:20 openhouse-spark-apps_2.12.jar
-rw-r--r--  1 anath1  101  201369249 Aug  2 10:21 openhouse-spark-apps_2.12-uber.jar
```

For all the boxes checked, include a detailed description of the testing done for the changes made in this pull request.

# Additional Information

- [x] Breaking Changes
- [ ] Deprecations
- [ ] Large PR broken into smaller PRs, and PR plan linked in the description.

For all the boxes checked, include additional details of the changes made in this pull request.